+++ app/src/main/java/com/univoice/feature/noticePost/NoticePostApplyFragment.kt
@@ -1,14 +1,11 @@
 package com.univoice.feature.noticePost
 
-import android.animation.Animator
-import android.content.Intent
 import android.view.View
 import androidx.navigation.NavOptions
 import androidx.navigation.fragment.findNavController
 import com.univoice.R
 import com.univoice.core_ui.base.BindingFragment
 import com.univoice.databinding.FragmentNoticePostApplyBinding
-import com.univoice.feature.MainActivity
 
 class NoticePostApplyFragment :
     BindingFragment<FragmentNoticePostApplyBinding>(R.layout.fragment_notice_post_apply) {
@@ -45,7 +42,11 @@ class NoticePostApplyFragment :
             val navOptions = NavOptions.Builder()
                 .setPopUpTo(findNavController().graph.startDestinationId, true)
                 .build()
-            findNavController().navigate(R.id.action_fragment_notice_post_apply_to_fragment_home, null, navOptions)
+            findNavController().navigate(
+                R.id.action_fragment_notice_post_apply_to_fragment_home,
+                null,
+                navOptions
+            )
         }
     }
