+++ app/src/main/java/com/univoice/feature/noticePost/timePicker/adapter/TimeDateAdapter.kt
@@ -1,19 +1,19 @@
-package com.univoice.feature.post.dateTimePicker.adapter
+package com.univoice.feature.noticePost.timePicker.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
 import androidx.recyclerview.widget.RecyclerView
 import com.univoice.R
+import com.univoice.feature.noticePost.timePicker.dpToPx
+import com.univoice.feature.noticePost.timePicker.getDateFromCalendar
 import java.util.Calendar
 
-
-class TimeHourAdapter(
-    val hour: ArrayList<Calendar>,
+class TimeDateAdapter(
+    val dates: ArrayList<Calendar>,
     private val dividerHeight: Int
 ) :
     RecyclerView.Adapter<RecyclerView.ViewHolder>() {
-
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         val view = DateViewHolder(
             DataBindingUtil.inflate(
@@ -31,11 +31,10 @@ class TimeHourAdapter(
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         val viewHolder = holder as DateViewHolder
 
-        viewHolder.binding.tvDateNumber.text = getHourFromCalendar(hour[position])
+        viewHolder.binding.tvDateNumber.text = getDateFromCalendar(dates[position])
     }
 
-
-    override fun getItemCount() = hour.size
+    override fun getItemCount() = dates.size
 
     override fun getItemId(position: Int): Long {
         return position.toLong()
