+++ app/src/main/res/drawable/bg_student_id_spinner.xml
